+++ R/filters.R
@@ -79,15 +79,17 @@
 #'
 #' - `locality` Character vector.
 #'   Filter by name of locality. Will first try to match strings to the
-#'   countries, provinces, and municipalities (see below) in FinBIF, if none of
-#'   these locality types match exactly then will return records with verbatim
-#'   locality matches in the original records.
+#'   countries, bio-provinces, and municipalities (see below) in FinBIF, if none
+#'   of these locality types match exactly then will return records with
+#'   verbatim locality matches in the original records.
 #' - `country` Character vector.
 #'   Filter by country. Use `finbif_metadata()` to see country names and ISO
 #'   codes (2 and 3 character) used in FinBIF.
-#' - `province` Character vector.
-#'   Filter by province. Use `finbif_metadata()` to see province names and
-#'   codes.
+#' - `region` Character vector.
+#'   Filter by region. Use `finbif_metadata()` to see region names and codes.
+#' - `bio_province` Character vector.
+#'   Filter by bio-province. Use `finbif_metadata()` to see bio-province names
+#'   and codes.
 #' - `municipality` Character vector.
 #'   Filter by municipality. Use `finbif_metadata()` to see municipality
 #'   names.