+++ data-raw/municipalities.R
@@ -23,5 +23,20 @@ names(municipality) <- c("name_en", "name_fi", "country")
 class(municipality[["name_en"]]) <- "translation"
 class(municipality[["name_fi"]]) <- "translation"
 
+file <- tempfile()
+geo_table <- download.file(
+  file.path(
+    "https://raw.githubusercontent.com",
+    "rOpenGov/geofi/master/data/municipality_key_2022.rda"
+  ),
+  destfile = file, quiet = TRUE
+)
+load(file)
+
+municipality[["region"]] <- as.data.frame(municipality_key_2022)[
+  order(municipality_key_2022[["municipality_name_fi"]]),
+  "maakunta_name_fi"
+]
+
 # english names are currently redundant
 municipality[["name_en"]] <- NULL